refactor(cache): split CacheWarmupService into partial-class files (no behavior change)
Mechanically extracts the 1982-LoC CacheWarmupService.cs into 8 partial-class
files grouped by responsibility. The compiled type, DI registration, field
initialization order, and public/protected method surface are unchanged —
partial classes compile to a single type with byte-identical semantics.

Motivation: 1k+ LoC files force code review (human and automated) to scan the
whole file plus several siblings to understand even small edits. Grouping
methods by responsibility lets reviewers load only the relevant subset.

File layout in src/Cache/Circles.Cache.Service/Services/:
- CacheWarmupService.cs (135 LoC) — fields, ctor, ClearCaches, nested types
- Warmup/CacheWarmupService.Lifecycle.cs (328) — ExecuteAsync, RunWarmupIterationAsync, DB/sync waits, timestamp helpers
- Warmup/CacheWarmupService.Connections.cs (60) — readonly + snapshot-bound connection helpers
- Warmup/CacheWarmupService.V1Events.cs (84) — V1 signup replay
- Warmup/CacheWarmupService.V2Events.cs (153) — V2 avatar/group/wrapper registration replay
- Warmup/CacheWarmupService.Balances.cs (714) — V1/V2/wrapper balance loaders + transfer fallbacks
- Warmup/CacheWarmupService.Metadata.cs (306) — group memberships, trust, consent flags, CIDs, short names
- Warmup/CacheWarmupService.GapReplay.cs (297) — block ring buffer init + post-Phase-2 gap replay

Verification (baseline captured before any edit, compared after split):
- Cache.Service build: 0 errors, 60 warnings (identical to baseline)
- Full repo build: 0 errors (clean)
- Cache.Service.Tests: 226 pass / 0 fail / 26 skip / 252 total (identical to baseline)
- Public method surface diff: empty (modulo expected partial-class declaration multiplication)

Out of scope: any behavior change, DI rewiring, IWarmupSchema strategy
extraction, or bug fix. The state-ownership refactor flagged by the 2026-03-04
audit is a separate concern and not addressed here.
